Partiion each table at different ratios

+++ benchmark_data_tools/generate_data_files.py
@@ -4,18 +4,20 @@ import json
 import subprocess
 import os
 import shutil
+import math
 
 from duckdb_utils import init_benchmark_tables, is_decimal_column
 from pathlib import Path
 from rewrite_parquet import process_dir
 from concurrent.futures import ThreadPoolExecutor
 
-def generate_partition(partition, raw_data_path, scale_factor, num_partitions, verbose):
+def generate_partition(table, partition, raw_data_path, scale_factor, num_partitions, verbose):
     if verbose:
-        print(f"Generating TPC-H partition: {partition}")
+        print(f"Generating '{table}' partition: {partition}")
     Path(f"{raw_data_path}/part-{partition}").mkdir(parents=True, exist_ok=True)
     command = [
         "tpchgen-cli",
+        "-T", table,
         "-s", str(scale_factor),
         "--output-dir", str(f"{raw_data_path}/part-{partition}"),
         "--parts", str(num_partitions),
@@ -27,22 +29,30 @@ def generate_partition(partition, raw_data_path, scale_factor, num_partitions, v
     except subprocess.CalledProcessError as e:
         print(f"Error generating TPC-H data: {e}")
 
-def generate_data_files_with_tpchgen(data_dir_path, scale_factor, convert_decimals_to_floats, num_partitions, num_threads, verbose):
-    if verbose:
-        print(f"Generating TPC-H data with {num_partitions} partitions")
+def generate_data_files_with_tpchgen(data_dir_path, scale_factor, convert_decimals_to_floats, num_threads, verbose):
 
     raw_data_path = data_dir_path + "-temp" if convert_decimals_to_floats else data_dir_path
+
+    # This dictionary maps each table to it's expected file size relative to the SF (roughly what SF is needed to generate a 5GB for this table).
+    # This is used to partition each table down to manageable file sizes.
+    tables_sf_ratio = {"region": 100000, "nation": 100000, "supplier": 100000, "customer": 200, "part": 100000, "partsupp": 100, "orders": 50, "lineitem": 10}
+
     with ThreadPoolExecutor(num_threads) as executor:
         futures = []
-        # Create partitioned data using tpchgen
-        for partition in range(0, num_partitions):
-            futures.append(executor.submit(generate_partition, partition, raw_data_path, scale_factor, num_partitions, verbose))
+
+        for table, ratio in tables_sf_ratio.items():
+            num_partitions = math.ceil(int(scale_factor) / ratio)
+            if verbose:
+                print(f"Generating TPC-H data for table '{table}' with {num_partitions} partitions")
+            for partition in range(1, num_partitions + 1):
+                futures.append(executor.submit(generate_partition, table, partition, raw_data_path, scale_factor, num_partitions, verbose))
+
         for future in futures:
             future.result()
 
     # When we generate partitioned data it will have the form <data_dir>/<partition>/<table_name>.parquet.
     # We want to re-arrange it to have the form <data_dir>/<table_name>/<table_name>-<partition>.parquet
-    parquet_files = os.listdir(f"{raw_data_path}/part-0")
+    parquet_files = os.listdir(f"{raw_data_path}/part-1")
     tables = []
     for p_file in parquet_files:
         tables.append(p_file.replace(".parquet", ""))
@@ -51,10 +61,11 @@ def generate_data_files_with_tpchgen(data_dir_path, scale_factor, convert_decima
         Path(f"{raw_data_path}/{table}").mkdir(parents=True, exist_ok=True)
 
     # Move the partitioned data into the new directory structure.
-    for partition in range(0, num_partitions):
+    for partition in range(1, num_partitions + 1):
         for table in tables:
-            shutil.move(f"{raw_data_path}/part-{partition}/{table}.parquet",
-                        f"{raw_data_path}/{table}/{table}-{partition}.parquet")
+            if os.path.exists(f"{raw_data_path}/part-{partition}/{table}.parquet"):
+                shutil.move(f"{raw_data_path}/part-{partition}/{table}.parquet",
+                            f"{raw_data_path}/{table}/{table}-{partition}.parquet")
         os.rmdir(f"{raw_data_path}/part-{partition}")
 
     if verbose:
@@ -85,15 +96,11 @@ def generate_data_files_with_duckdb(benchmark_type, data_dir_path, scale_factor,
 def generate_data_files(benchmark_type, data_dir_path, scale_factor, convert_decimals_to_floats, use_duckdb, num_threads, verbose):
     Path(f"{data_dir_path}").mkdir(parents=True, exist_ok=True)
     # tpchgen is much faster, but is exclusive to generating tpch data.  Use duckdb as a fallback.
-    if benchmark_type == "tpch" and not use_duckdb:
-        try:
-            int_scale_factor = int(scale_factor)
-            num_partitions = int(int_scale_factor / 10) if int_scale_factor >= 10 else int(1)
-            generate_data_files_with_tpchgen(data_dir_path, scale_factor, convert_decimals_to_floats, num_partitions, num_threads, verbose)
-            return
-        except ValueError:
-            print(f"scale factor {scale_factor} too small to use tpchgen; falling back to duckdb")
-    generate_data_files_with_duckdb(benchmark_type, data_dir_path, scale_factor, convert_decimals_to_floats)
+    if benchmark_type == "tpch" and not use_duckdb and float(scale_factor) >= 1:
+        generate_data_files_with_tpchgen(data_dir_path, scale_factor, convert_decimals_to_floats, num_threads, verbose)
+    else:
+        print(f"scale factor {scale_factor} too small to use tpchgen; falling back to duckdb")
+        generate_data_files_with_duckdb(benchmark_type, data_dir_path, scale_factor, convert_decimals_to_floats)
 
 def get_select_query(table_name, convert_decimals_to_floats):
     if convert_decimals_to_floats:
